pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thinqconnect-mcp"
version = "0.0.4"
description = "ThinQ Connect MCP"
authors = [
    {name = "ThinQ Connect", email = "thinq-connect@lge.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "thinqconnect",
    "python-dotenv",
    "mcp",
    "aiohttp"
]

[project.scripts]
thinqconnect-mcp = "thinqconnect_mcp.server:main"

[project.urls]
Homepage = "https://github.com/thinq-connect/thinqconnect-mcp"

[tool.setuptools.packages.find]
where = ["."]
include = ["thinqconnect_mcp*"]
